支持直接存储与读取URL, 直接存储URL的不支持读取content

+++ qcloudcos/qcloudstorage.py
@@ -13,11 +13,17 @@ class QcloudStorage(Storage):
             self.option = settings.QCLOUD_STORAGE_OPTION
 
     def _open(self, name, mode='rb'):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return ''
         cos = CosObject()
         response = cos.get_object(name, True)
         return response.content
 
     def _save(self, name, content):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return name
         name = self._get_valid_name(name)
         name = self._get_available_name(name)
         content = content.read()
@@ -26,6 +32,9 @@ class QcloudStorage(Storage):
         return response.request.path_url
 
     def _get_valid_name(self, name):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return name
         dir_name, file_name = os.path.split(name)
         file_name = get_valid_filename(file_name)
         name = os.path.join(dir_name, file_name)
@@ -54,6 +63,9 @@ class QcloudStorage(Storage):
         return name
 
     def exists(self, name):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return True
         name = self._get_valid_name(name)
         cos = CosObject()
         response = cos.head_object(name, True)
@@ -63,6 +75,9 @@ class QcloudStorage(Storage):
             return False
 
     def url(self, name):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return name
         if settings.COS_USE_CDN:
             cdn_host = 'file'
         else:
@@ -77,6 +92,9 @@ class QcloudStorage(Storage):
         return url
 
     def size(self, name):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return 0
         name = self._get_valid_name(name)
         cos = CosObject()
         response = cos.head_object(name, True)
@@ -84,6 +102,9 @@ class QcloudStorage(Storage):
             return response.headers['Content-Length']
 
     def delete(self, name):
+        if name.startswith('http'):
+            # 直接存的URL，直接返回，这类数据不支持取content
+            return
         name = self._get_valid_name(name)
         cos = CosObject()
         cos.delete_object(name)
